+++ test/controller/test/sub/sub2/foo.js
@@ -1,3 +1,8 @@
+var gn = require('gracenode');
+var logger = gn.log.create();
+
 exports.GET = function (req, res) {
+	logger.debug('controller:', req.controller);
+	logger.debug('method:', req.method);
 	res.json({ method: 'sub2/foo', params: req.parameters });
 };