+++ test/controller/test/sub/sub2/foo.js
@@ -4,5 +4,5 @@ var logger = gn.log.create();
 exports.GET = function (req, res) {
 	logger.debug('controller:', req.controller);
 	logger.debug('method:', req.method);
-	res.json({ method: 'sub2/foo', params: req.parameters });
+	res.json({ key: req.get('key'), method: 'sub2/foo', params: req.parameters });
 };